requirements.txt

Flask==1.1.2
Faker==9.9.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
snowflake-connector-python==2.6.2
simplejson==3.17.6
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
SQLAlchemy-serializer==1.4.1
